// src/types/station_data.ts
export interface StationData {
  STOP_ID: number;
  DIRECTION_ID: string;
  STOP_NAME: string;
  STATION_NAME: string;
  STATION_DESCRIPTIVE_NAME: string;
  MAP_ID: number;
  ADA: boolean;
  RED: boolean;
  BLUE: boolean;
  G: boolean;
  BRN: boolean;
  P: boolean;
  Pexp: boolean;
  Y: boolean;
  Pnk: boolean;
  O: boolean;
  Location: string;
}